+++ Simulation.java
@@ -5,8 +5,8 @@ import grabl.tracing.client.GrablTracingFactory;
 import grakn.client.GraknClient;
 import grakn.client.GraknClient.Session;
 import grakn.client.GraknClient.Transaction;
-import grakn.simulation.agents.AgentContext;
-import grakn.simulation.agents.AgentRunner;
+import grakn.simulation.agents.base.AgentContext;
+import grakn.simulation.agents.base.AgentRunner;
 import grakn.simulation.agents.World;
 import grakn.simulation.common.RandomSource;
 import grakn.simulation.yaml_tool.GraknYAMLException;
