+++ agents/base/AgentRunner.java
@@ -1,10 +1,9 @@
-package grakn.simulation.agents;
+package grakn.simulation.agents.base;
 
 import grakn.simulation.common.Pair;
 import grakn.simulation.common.RandomSource;
 
 import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.Random;
